+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Added
 - Annotation `@phpstan-latte-ignore` can be used to ignore render calls, variable assignments, component creation, whole methods or classes.
 - Annotation `@phpstan-latte-template` can be used to specify what template is used to render.
-- Annotation `@phpstan-latte-var` can be used to specify what variables are available in tempalte.
+- Annotation `@phpstan-latte-var` can be used to specify what variables are available in template.
+- Annotation `@phpstan-latte-component` can be used to specify what components are available in template.
 - Resolves calls to `setView` and `sendTemplate` in presenters
  
 ## [0.2.0] - 2022-12-06